Lots of changes. See Description
Edited scene switching so that scenes are switched by placing a marble in a cup. Scenes available include sandbox and level 1. There is a quit option and in each scene there is an option to return to the menu.
Added Recycle bin to level 1. It doesnt quite work properly because each itemed spawned has a different name.