Added the scoring scripts
the scoring script keeps track of the time since the level loaded as well as each time the marble touches a track piece. This is under the variables time passed and tracks used. The marble scored and conditions met dont update yet when the goal is reached. However, when the marble reaches the goal, the position is locked and the reset button no longer functions.