Added a few changes
Changes: Edited the text walls, Changed selector material in level 1. Edited placement of level 1 start and end. Added text mesh pro. The problem with the current text is that it renders in front of any object, even if the object is physically in front of it. TextMeshPro should help. To compensate for this, i have moved text around a bit so that it does not get in the way of playing.