Level 3 Added
Added a 3rd level that uses a portal mechanic. The user has to build a ramp that gets the marble going a certain speed through the portal. If the speed is enough, it will travel from the exit portal on the robots hand, and into the in the robots other hand.